feat(rtic-v2-rtc-monotonics): Further prepares the monotonics for publication in the HAL.
* Moves the two monotonics into their own files so that the `rtc::rtic` main module file is not so large and unwieldy.
* Adds thorough documentation to the `rtc::rtic` module and its items, discussing the differences between the two monotonics, and how to use the macros to create them.
